Correctly convert arguments.
